feat: add deg and rad funcs
When working with trignometric functions, it is useful to be able to
quickly convert values between radians and degrees, and vice versa.

The function `deg` converts its input to degrees, and rad converts its
input to degrees.

An example of the utility of `deg` and `rad` can be demonstrated with
the following example:

```
> deg(asin(1))
90
> sin(rad(90))
1
```