feat: add a sqrt func
When an expression such as `2^1/2` is entered, it is parsed as `(2^1)/2`
rather than `2^(1/2)` so square rooting numbers, which is a common
operation requires the use of a set of parenthesis.

Instead, the `sqrt` function is provided as a way to conveniently square
root numbers; `sqrt(x)` it is equivalent to `(x^(1/2))`.